+++ AIPScannerConfig/functions/Add-AIPRerpository.ps1
@@ -39,7 +39,7 @@
         None
     #>
 
-    [CmdletBinding()]
+    [CmdletBinding(PSUseShouldProcessForStateChangingFunctions = $true)]
     [OutputType([System.Boolean])]
     param (
         [Parameter(Position = 0, ParameterSetName = 'FileShare', HelpMessage = 'Path to file share')]