+++ AIPScannerConfig/internal/functions/Assert-ElevatedPermission.ps1
@@ -1,4 +1,4 @@
-﻿function Assert-ElevatedPermissions {
+﻿function Assert-ElevatedPermission {
     <#
     .SYNOPSIS
         Check the elevation level
@@ -6,9 +6,6 @@
     .DESCRIPTION
         Checks to see if we are running PowerShell with Administrator credentials
     
-    .EXAMPLE
-        PS C:\> Assert-ElevatedPermissions
-
     .NOTES
         Internal function
         Exists out if the interactive PowerShell session is not being run with administrative credentials
@@ -19,15 +16,15 @@
     param ()
     
     begin {
-        Write-PSFMessage -Level Host -String 'Assert-ElevatedPermissions.Message1'
+        Write-PSFMessage -Level Host -String 'Assert-ElevatedPermission.Message1'
     }
     process{
         if (-NOT ([Security.Principal.WindowsPrincipal][Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltInRole] "Administrator")) {
-            Write-PSFMessage -Level Host -String 'Assert-ElevatedPermissions.Message2' 
+            Write-PSFMessage -Level Host -String 'Assert-ElevatedPermission.Message2'
             return $False
         }
         else {
-            Write-PSFMessage -Level Host -String 'Assert-ElevatedPermissions.Message3'
+            Write-PSFMessage -Level Host -String 'Assert-ElevatedPermission.Message3'
             return $True
         }
     }